[MISC][RL] Rename file to match convention
