+++ src/form/index.ts
@@ -4,12 +4,12 @@ import { FormInput } from "./form-input";
 import { FormInputGroup } from "./form-input-group";
 import { FormLabel } from "./form-label";
 import { FormMultiSelect } from "./form-multi-select";
+import { FormPhoneNumberInput } from "./form-phone-number-input";
 import { FormSelect } from "./form-select";
 import { FormTextarea } from "./form-textarea";
+import { FormTimeRangePicker } from "./form-time-range-picker";
 import { FormTimepicker } from "./form-timepicker";
-import { FormTimeRangePicker } from "./form-timerangepicker";
 import { FormUnitNumberInput } from "./form-unit-number-input";
-import { FormPhoneNumberInput } from "./form-phone-number-input";
 
 export const Form = {
     DateInput: FormDateInput,
