+++ src/index.ts
@@ -38,11 +38,11 @@ export * from "./smart-app-banner";
 export * from "./text";
 export * from "./text-list";
 export * from "./theme";
+export * from "./time-range-picker";
 export * from "./time-slot-bar";
 export * from "./timeline";
 export * from "./timepicker";
-export * from "./time-range-picker";
+export * from "./toast";
 export * from "./toggle";
 export * from "./tooltip";
 export * from "./transition";
-export * from "./toast";
